neutrino: Introduce NeutrinoSettings
Add support for an alternative approach to systemless ramdisk
modification by performing most of the heavy lifting with
NeutrinoSettings, a built-in Magisk module that is installed
alongside the kernel when root is detected. This is preferable to
ramdisk overlays as it allows the files to be accessed while booted
and removes the reliance on subsequent kernel installations to
clean up after us. A secondary late-init script is run from
service.d on boot to determine if NeutrinoKernel is still installed
and initialize the module. If another kernel is detected the module
is auto-removed before the script self-destructs. All traces of
Neutrino will be removed following a subsequent reboot without the
need for external intervention.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>