neutrino: anykernel.sh: Support both OOS and custom ROMs in the same zip
We can build both images and include them in the zip then
dynamically move them to where they need to go.

This is an extension of what we used to do with OnePlus6T (0ctobot/AnyKernel3@da88709)

However, things have gotten a little more complicated this time
around with the inordinate amount of device variants and codenames
that we're supporting here. Furthermore, OnePlus7(Pro) device
build.props contain two instances of ro.build.user/ro.build.flavor
each, thus we've included a secondary file_getprop function to
accommodate this unusual scenario and ensure that we're returning
the desired match when handling those devices.

[@0ctobot: Adapted from nathanchance/AnyKernel2-archive@bfa82cf]
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>